Added sounds for Animals -> Land
Added tags (0-2) for land animal buttons.

Imported Sounds folder into Xcode.

Set the class for the View controller regarding the land animal sounds to "ViewController" in the Assisant Editor.

Drag a sound button from land animals into the viewController class (code).
Set the button type to UIButton and called the function soundPressed.

Dragged the remaining 2 sound buttons onto the function.

Importanted AVFoundation.

Created a soundList array containing all the sounds, Soundn
Created a soundName string

Checked if the button tag is less than the array size and if the button tag is greater tthan or equal to 0
Set soundName to the correct sound in the array based of the button tag.

Played the sound.