Added the remaing sounds & function to pause when switch tab
Addeed the remaing sounds for the app using the same logic in the previous commits.

Connected all sounds to be able to be toggled on/off.

Added the function for the sound to pause when switching to another tab, so that two sounds, one from each tab, cannot be simultaneously played.